requirements.txt
fastapi
uvicorn
sqlalchemy
mysqlclient
pydantic
python-dotenv